Separated multer configuration for upload into separate file
