+++ lib/server.js
@@ -3,22 +3,7 @@ const network = require('network');
 const debug = require('debug')('qrf:server');
 const zip = require('express-easy-zip');
 const app = express();
-
-const multer = require('multer');
-let path = require("path");
-
-const storage = multer.diskStorage({
-  destination: function(req, file, cb) {
-    cb(null,path.join(__dirname+'/uploads'));
-  },
-  filename: function(req, file, cb) {
-    cb(null, file.originalname);
-  }
-});
-
-const upload = multer({ 
-  storage: storage 
-});
+const upload = require('./upload-middleware');
 
 const {
   FILE,
